% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{glasspockets.api}
\alias{glasspockets.api}
\title{Submits a search query to the glasspockets.org API, saves the result as JSON,
and returns a list object representing the JSON contents}
\usage{
glasspockets.api(query, file.name, api.key)
}
\arguments{
\item{query}{String search query, typically a foundation name}

\item{file.name}{String file name (can include path) to save the JSON file as}

\item{api.key}{An API key generated from glasspockets.org}
}
\value{
Returns a dataframe where every row is a grant
}
\description{
Submits a search query to the glasspockets.org API, saves the result as JSON,
and returns a list object representing the JSON contents
}
\examples{
# setup your parameters
api.key <- "YOUR_API_KEY"
file.name <- "gates_foundation.json"
query <- "Gates"

# submit your query
gates.grants <- glasspockets.api(query, file.name, api.key)
}

