% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{load.grants}
\alias{load.grants}
\title{Extracts grants from the JSON list returned by the glasspockets::glasspockets(...) method.
Also converts date strings to date objects.}
\usage{
load.grants(file.name)
}
\arguments{
\item{file.name}{File name (or path plus name) of the JSON file to load}
}
\value{
Returns a dataframe where each row is a grant.
}
\description{
Extracts grants from the JSON list returned by the glasspockets::glasspockets(...) method.
Also converts date strings to date objects.
}
\examples{
gates.grants <- load.grants("gates.json")
}

